requirements.txt
ansi2html==1.8.0
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
dash==2.13.0
dash-bootstrap-components==1.5.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
distro==1.8.0
et-xmlfile==1.1.0
Flask==2.2.5
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
nest-asyncio==1.5.8
numpy==1.26.0
openpyxl==3.1.2
pandas==2.1.1
pdfminer.six==20200517
pdfplumber==0.5.28
Pillow==10.0.1
plotly==5.3.1
pycryptodome==3.19.0
PyPDF2==1.26.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
retrying==1.3.4
scipy==1.11.3
six==1.16.0
sortedcontainers==2.4.0
tabula-py==2.3.0
tenacity==8.2.3
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
Wand==0.6.11
Werkzeug==2.2.3
